requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
python-dotenv
python-multipart
Pillow
tensorflow